add webclient classes,  ClientResponseDto, some services and util classes

+++ src/main/java/com/iivanov/cleverdevtestnewsystem/entities/Patient.java
@@ -27,7 +27,7 @@ public class Patient extends AbstractEntity {
     private String oldClientGuid;
 
     @Column(name = "status_id", nullable = false)
-    private Integer status;
+    private short status;
 
     @Override
     public boolean equals(Object o) {
